Add error handling to DataStore lpa repository
